Convert all js files to typescript and use ts-node.
