+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.3.4",
   "private": true,
   "scripts": {
-    "start": "node ./src/botservice.js"
+    "start": "ts-node ./src/botservice.ts"
   },
   "dependencies": {
     "@mattermost/client": "^7.8.0",
@@ -12,7 +12,12 @@
     "form-data": "^4.0.0",
     "isomorphic-fetch": "^3.0.0",
     "openai": "^3.2.1",
-    "ws": "^8.12.1"
+    "ws": "^8.12.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.5"
+  },
+  "devDependencies": {
+    "@types/ws": "^8.5.5"
   },
   "engines": {
     "npm": ">=8.0.0",
