+++ tsconfig.json
@@ -1,0 +1,22 @@
+{
+  "extends": "ts-node/node16/tsconfig.json",
+  "compilerOptions": {
+    "module": "ESNext",
+    "target": "ESNext",
+    "sourceMap": true,
+    "rootDir": "src",
+    "moduleResolution": "node",
+    "lib": ["DOM"],
+    "resolveJsonModule": true,
+    "esModuleInterop": true
+  },
+  "ts-node": {
+    "esm": true,
+    "compilerOptions": {
+      "module": "NodeNext",
+    }
+  },
+  "include": ["src"],
+
+  "exclude": ["node_modules"]
+}
